requirements.txt
streamlit
streamlit-webrtc
av
opencv-python
azure-ai-vision-imageanalysis
azure-core
python-dotenv